requirements.txt
boto3~=1.28.35
config
Flask~=2.3.3
Flask-Login
Flask-SQLAlchemy
folium~=0.14.0
gunicorn
numpy
pandas
plotly
plotly-express
python-dotenv~=1.0.0
pymongo
Werkzeug~=2.3.7
opencv-python-headless


SQLAlchemy~=2.0.20